# book-network-ui/src/app/services/models/api-response-authentication-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AuthenticationResponse } from './authentication-response';
export interface ApiResponseAuthenticationResponse {
  data?: AuthenticationResponse;
  message?: string;
}