# book-network-ui/src/app/services/models/api-response-register-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RegisterResponse } from './register-response';
export interface ApiResponseRegisterResponse {
  data?: RegisterResponse;
  message?: string;
}